requirements_docker.txt
click==8.1.7
hydra-core==1.3.2
matplotlib==3.8.2
numpy==1.24.1
omegaconf==2.3.0
pytorch_lightning==2.1.3
scikit_learn==1.3.2
tqdm==4.66.1
wget==3.2
